fix(fill): preserve fractional rectilinear offsets

+++ crates/ares-core/src/fill/rectilinear/surface.rs
@@ -103,11 +103,11 @@ pub(super) fn scaled_offsets(
 }
 
 fn checked_scaled_f32(scale: CoordinateScale, value: f64) -> Result<f32, ClipperError> {
-    let scaled = (value / scale.factor()) as f32;
-    if !scaled.is_finite() || !(i64::MIN as f32..-(i64::MIN as f32)).contains(&scaled) {
+    let scaled = value / scale.factor();
+    if !scaled.is_finite() || !(i64::MIN as f64..-(i64::MIN as f64)).contains(&scaled) {
         return Err(ClipperError::CoordinateOutOfRange);
     }
-    Ok((scaled as i64) as f32)
+    Ok(scaled as f32)
 }
 
 fn adjust_solid_spacing(width: i64, distance: i64) -> i64 {
